move the currency module to folder
